Authenticated users can apply any license
And DRY up the media items form partial.
