+++ app/controllers/application_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class ApplicationController < ActionController::Base
+  include CurrentUserHelper
+
   protect_from_forgery with: :exception
   around_action :set_current_user
 
@@ -21,24 +23,7 @@ class ApplicationController < ActionController::Base
 
   private
 
-  def current_user_ability
-    @current_user_ability ||= Ability.new(current_user)
-  end
-
-  def current_user_can?(*args)
-    current_user_ability.can?(*args)
-  end
-
   def render_http_status(status)
     render plain: Rack::Utils::HTTP_STATUS_CODES[status], status: status
-  end
-
-  def set_current_user
-    Current.user = current_user
-    yield
-  ensure
-    # to address the thread variable leak issues in Puma/Thin webserver,
-    # see https://stackoverflow.com/a/8291218/738371
-    Current.user = nil
   end
 end
