+++ app/views/campaigns/generic/field_sets/_media_items.html.erb
@@ -1,81 +1,26 @@
 <%= field_set_tag nil, class: 'media-items' do %>
   <div id="media-item-edm_isShownBy" class="nested_fields">
-    <%= field_set_tag do %>
-      <%= agg.simple_fields_for(:edm_isShownBy) do |wr| %>
-        <%= wr.input(:media,
-                     label: campaigning.t('legends.each_object', number: 1),
-                     label_html: { class: 'sequenced_object' },
-                     hint: campaigning.t('hints.edm_hasView.media'),
-                     input_html: { accept: "#{EDM::WebResource.allowed_extensions}, #{EDM::WebResource.allowed_content_types}", 'data-makes-required': '#media-item-edm_isShownBy .license-radios',
-                     'data-max-bytes': EDM::WebResource.max_media_size })
-        %>
-        <%= wr.input(:media_cache, as: :hidden, 'data-makes-required': '#media-item-edm_isShownBy .license-radios') %>
-        <% if wr.object.media? %>
-          <div class="cf">
-            <div class="media-preview">
-              <%= image_tag(wr.object.media_url) %>
-              <%= wr.input(:remove_media, as: :boolean) %>
-            </div>
-          </div>
-        <% end %>
-        <%= wr.input(:dc_description) %>
-        <%= wr.input(:dc_type, as: :autocomplete, url: vocabularies_europeana_contribute_getty_aat_path, param: 'q', input_html: { 'data-deref-url': vocabularies_europeana_contribute_getty_aat_dereference_path }) %>
-        <%= wr.input(:dcterms_created, as: :date, html5: true, input_html: date_format_fallback_attributes) %>
-        <%= wr.input(:dc_creator) %>
-        <%= wr.input(:edm_rights_id,
-                     as: :radio_buttons,
-                     collection: edm_rights_options,
-                     item_wrapper_class: 'license-radio-option',
-                     collection_wrapper_class: 'licenses',
-                     collection_wrapper_tag: 'div',
-                     item_wrapper_tag: 'span',
-                     wrapper_html: { class: 'license-radios', 'data-error-msg-key': 'contribute.campaigns.generic.form.validation.web-resource-license' },
-                     label_input_html: { 'class': 'license-section' })
-        %>
-      <% end %>
+    <%= agg.simple_fields_for(:edm_isShownBy) do |web_resource| %>
+      <%= render partial: '/campaigns/generic/field_sets/web_resource', locals: {
+        web_resource: web_resource, number: 1, removable?: false, campaigning: campaigning,
+        media_input_makes_required: '#media-item-edm_isShownBy .license-radios',
+        media_cache_input_makes_required: '#media-item-edm_isShownBy .license-radios'
+      } %>
     <% end %>
   </div>
 
-  <%= agg.nested_fields_for(:edm_hasViews, wrapper_tag: :div, wrapper_options: { id: 'media-item-edm_hasView-__nested_field_for_replace_with_index__' }) do |wr| %>
-    <%= field_set_tag do %>
-      <%= wr.remove_nested_fields_link(campaigning.t('buttons.object.remove')) %>
-      <% number = if agg.object.edm_hasViews.include?(wr.object)
-                    agg.object.edm_hasViews.index(wr.object) + 2
-                  else
-                    agg.object.edm_web_resources.size + 1
-                  end
-      %>
-      <%= wr.input(:media,
-                   label: campaigning.t('legends.each_object', number: number),
-                   label_html: { class: 'sequenced_object' },
-                   hint: campaigning.t('hints.edm_hasView.media'),
-                   input_html: { accept: "#{EDM::WebResource.allowed_extensions}, #{EDM::WebResource.allowed_content_types}", 'data-makes-required': '#media-item-edm_hasView-__nested_field_for_replace_with_index__ .license-radios',
-                   'data-max-bytes': EDM::WebResource.max_media_size })
-      %>
-      <%= wr.input(:media_cache, as: :hidden, 'data-makes-required': '#media-item-edm_hasView-__nested_field_for_replace_with_index__ .license-radios') %>
-      <% if wr.object.media? %>
-        <div class="cf">
-          <div class="media-preview">
-            <%= image_tag(wr.object.media_url) %>
-            <%= wr.input(:remove_media, as: :boolean) %>
-          </div>
-        </div>
-      <% end %>
-      <%= wr.input(:dc_description) %>
-      <%= wr.input(:dc_type, as: :autocomplete, url: vocabularies_europeana_contribute_getty_aat_path, param: 'q', input_html: { 'data-deref-url': vocabularies_europeana_contribute_getty_aat_dereference_path }) %>
-      <%= wr.input(:dcterms_created, as: :date, html5: true, input_html: date_format_fallback_attributes) %>
-      <%= wr.input(:dc_creator) %>
-      <%= wr.input(:edm_rights_id,
-                   as: :radio_buttons,
-                   collection: edm_rights_options,
-                   item_wrapper_class: 'license-radio-option',
-                   collection_wrapper_class: 'licenses',
-                   collection_wrapper_tag: 'div',
-                   item_wrapper_tag: 'span',
-                   wrapper_html: { class: 'license-radios', 'data-error-msg-key': 'contribute.campaigns.generic.form.validation.web-resource-license' },
-                   label_input_html: { 'class': 'license-section' })
-      %>
-    <% end %>
+  <%= agg.nested_fields_for(:edm_hasViews, wrapper_tag: :div, wrapper_options: { id: 'media-item-edm_hasView-__nested_field_for_replace_with_index__' }) do |web_resource| %>
+    <% number = if agg.object.edm_hasViews.include?(web_resource.object)
+                  agg.object.edm_hasViews.index(web_resource.object) + 2
+                else
+                  agg.object.edm_web_resources.size + 1
+                end
+    %>
+    <%= render partial: '/campaigns/generic/field_sets/web_resource', locals: {
+      web_resource: web_resource, number: number, removable?: true, campaigning: campaigning,
+      media_input_makes_required: '#media-item-edm_hasView-__nested_field_for_replace_with_index__ .license-radios',
+      media_cache_input_makes_required: '#media-item-edm_hasView-__nested_field_for_replace_with_index__ .license-radios'
+    } %>
   <% end %>
   <%= agg.add_nested_fields_link(:edm_hasViews, campaigning.t('buttons.object.add')) %>
 <% end %>
